+++ fees/ramses-exchange-v2.ts
@@ -1,0 +1,63 @@
+import { SimpleAdapter, FetchResultFees, BaseAdapter } from "../adapters/types";
+import { ARBITRUM, CHAIN } from "../helpers/chains";
+
+
+import {
+    getGraphDimensions,
+    DEFAULT_DAILY_VOLUME_FACTORY,
+    DEFAULT_TOTAL_VOLUME_FIELD,
+  } from "../helpers/getUniSubgraph"
+
+type TStartTime = {
+[key: string]: number;
+}
+const startTimeV2:TStartTime = {
+[CHAIN.ARBITRUM]: 1685574000,
+}
+  
+
+const v2Endpoints = {
+    [CHAIN.ARBITRUM]: "https://api.thegraph.com/subgraphs/name/ramsesexchange/concentrated-liquidity-graph",
+  };
+  
+const VOLUME_USD = "volumeUSD";
+
+const v2Graphs = getGraphDimensions({
+    graphUrls: v2Endpoints,
+    totalVolume: {
+        factory: "factories",
+        field: DEFAULT_TOTAL_VOLUME_FIELD,
+    },
+    dailyVolume: {
+        factory: DEFAULT_DAILY_VOLUME_FACTORY,
+        field: VOLUME_USD,
+    },
+    feesPercent: {
+        type: "fees",
+        HoldersRevenue: 100,
+        UserFees: 100, // User fees are 100% of collected fees
+        Revenue: 100 // Revenue is 100% of collected fees
+    }
+    });
+
+    const methodology = {
+    UserFees: "User pays 0.3% fees on each swap.",
+    ProtocolRevenue: "Revenue going to the protocol.",
+    HoldersRevenue: "User fees are distributed among holders."
+    }
+
+    const adapter: SimpleAdapter = {
+        adapter: {
+          [CHAIN.ARBITRUM]: {fetch: v2Graphs(ARBITRUM),
+            start: async () => startTimeV2[CHAIN.ARBITRUM],
+            meta: {
+              methodology: {
+                ...methodology,
+                UserFees: "User pays 0.05%, 0.30%, or 1% on each swap."
+              }
+            }
+          },
+        }
+      };
+
+export default adapter;
